fix: 중복된 패키지와 클래스 제거
- Controller, Service, Repository 중복 제거
- CategoryController 분리
