--- our-memory/src/main/java/keepsake/ourmemory/repository/MemoryRepository.java
@@ -1,8 +1,0 @@
-package keepsake.ourmemory.repository;
-
-import keepsake.ourmemory.domain.memory.Memory;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemoryRepository extends JpaRepository<Memory, Long> {
-
-}
