+++ our-memory/src/main/java/keepsake/ourmemory/ui/CategoryController.java
@@ -1,34 +1,25 @@
-package keepsake.ourmemory.api.memory;
+package keepsake.ourmemory.ui;
 
-import jakarta.validation.Valid;
-import keepsake.ourmemory.api.memory.response.CategoriesResponse;
-import keepsake.ourmemory.api.memory.response.CategoryResponse;
 import keepsake.ourmemory.application.memory.MemoryService;
 import keepsake.ourmemory.application.memory.dto.CategoryDto;
+import keepsake.ourmemory.ui.dto.response.CategoriesResponse;
+import keepsake.ourmemory.ui.dto.response.CategoryResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.net.URI;
 import java.util.List;
 
-@RequiredArgsConstructor
 @RestController
-public class MemoryController {
-
+@RequestMapping("/categories")
+@RequiredArgsConstructor
+public class CategoryController {
     private final MemoryService memoryService;
 
-    @PostMapping("/memories")
-    public ResponseEntity<Void> createMemory(@Valid @RequestBody MemoryCreateRequest request) {
-        Long memoryId = memoryService.createMemory(request);
-        return ResponseEntity.created(URI.create("/memories/" + memoryId)).build();
-    }
-
-    @GetMapping("/categories")
+    @GetMapping
     public ResponseEntity<CategoriesResponse> findCategories() {
         List<CategoryDto> categories = memoryService.findCategories();
         CategoriesResponse response = new CategoriesResponse(
