--- our-memory/src/main/java/keepsake/ourmemory/ui/dto/TagResponse.java
@@ -1,4 +1,0 @@
-package keepsake.ourmemory.ui.dto;
-
-public record TagResponse(Long id, String name, String color) {
-}
