+++ our-memory/src/main/java/keepsake/ourmemory/ui/dto/response/CategoryResponse.java
@@ -1,4 +1,4 @@
-package keepsake.ourmemory.api.memory.response;
+package keepsake.ourmemory.ui.dto.response;
 
 import keepsake.ourmemory.application.memory.dto.CategoryDto;
