+++ our-memory/src/main/java/keepsake/ourmemory/ui/dto/response/LocationResponse.java
@@ -1,4 +1,4 @@
-package keepsake.ourmemory.ui.dto;
+package keepsake.ourmemory.ui.dto.response;
 
 public record LocationResponse(String latitude, String longitude) {
 }
