+++ our-memory/src/main/java/keepsake/ourmemory/ui/dto/response/MemoriesResponse.java
@@ -1,4 +1,4 @@
-package keepsake.ourmemory.ui.dto;
+package keepsake.ourmemory.ui.dto.response;
 
 
 import java.util.List;
