+++ our-memory/src/main/resources/application.yml
@@ -11,7 +11,5 @@ spring:
       hibernate:
 #        show_sql: true
         format_sql: true
+        highlight_sql: true
         default_batch_fetch_size: 100 #최적화 옵션
-
-logging.level:
-  org.hibernate.SQL: debug
