+++ our-memory/src/test/java/keepsake/ourmemory/application/MemoryServiceTest.java
@@ -1,11 +1,11 @@
 package keepsake.ourmemory.application;
 
-import keepsake.ourmemory.api.memory.MemoryCreateRequest;
+import keepsake.ourmemory.application.repository.MemoryRepository;
+import keepsake.ourmemory.ui.dto.request.MemoryCreateRequest;
 import keepsake.ourmemory.application.memory.MemoryService;
 import keepsake.ourmemory.domain.memory.Category;
 import keepsake.ourmemory.domain.memory.Memory;
 import keepsake.ourmemory.domain.memory.Star;
-import keepsake.ourmemory.repository.MemoryRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.ArgumentMatchers;
